% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePerformances.R
\name{ComputePerformances}
\alias{ComputePerformances}
\title{Compute the performances of the l1-spectral clustering algorithm}
\usage{
ComputePerformances(Results, A)
}
\arguments{
\item{Results}{Output of the function \code{l1_spectralclustering()}.}

\item{A}{The adjacency matrix of the graph to cluster.}
}
\value{
The Normalized Mutualized Information (NMI) score.
}
\description{
This function computes the performances of the l1-spectral clustering algorithm in terms of Normalized Mutualized Information (NMI).
}
\examples{
 #############################################################
 # Generating toy data
 #############################################################
 Data <- CreateDataSet(k=3, n=20, p=list(p_inside=0.1,p_outside=0.1))

}
\seealso{
\code{\link{l1_spectralclustering}}, \code{\link{l1spectral}}.
}
\author{
Camille Champion, Magali Champion
}
